+++ Kernel/cpan-lib/MIME/Decoder/NBit.pm
@@ -98,7 +98,7 @@ use MIME::Tools qw(:msgs);
 @ISA = qw(MIME::Decoder);
 
 ### The package version, both in 1.23 style *and* usable by MakeMaker:
-$VERSION = "5.506";
+$VERSION = "5.507";
 
 ### How many bytes to decode at a time?
 my $DecodeChunkLength = 8 * 1024;