+++ Kernel/cpan-lib/MIME/Head.pm
@@ -138,7 +138,7 @@ use MIME::Field::ContType;
 #------------------------------
 
 ### The package version, both in 1.23 style *and* usable by MakeMaker:
-$VERSION = "5.506";
+$VERSION = "5.507";
 
 ### Sanity (we put this test after our own version, for CPAN::):
 use Mail::Header 1.06 ();